+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Added `up` command to navigate to parent branch in the stack
+- Added `down` command to navigate to child branches with smart selection for multiple children
+
 ## [v0.11.0] - 2025-05-13
 
 ### Added